% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_functions.R
\name{build.overwinter}
\alias{build.overwinter}
\title{Raster-based Overwintering calculations}
\usage{
build.overwinter(brick_name = Ta_range, t.thresh = 18, n.days = 5)
}
\arguments{
\item{brick_name}{The \code{RasterBrick} object e.g. Tmin}

\item{t.thresh}{Threshold temperature in degrees}

\item{n.days}{Consecutive days needed at threshold to change condition}
}
\value{
\code{RasterLayer} object showing total number of overwintering days for the period analysed
}
\description{
Similar to \code{overwinter()} but spatial. Calculates number of overwintering days:
these are triggered by X consecutive days below a threshold temperature / broken by X days above
}
\author{
Matt Hill
}
