% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_functions.R
\name{cherry_dev}
\alias{cherry_dev}
\title{Simple cherry_dev}
\usage{
cherry_dev(tlow = 4.5, thigh = 100, tempvec = tempvec,
  budburst = "2015-08-31")
}
\description{
Switch for overwintering. This calculates cumulative days less than 18 degrees (5 needed to enter)
Currently calculated on the Ta (Average temperature for the day - based on the interpolated temp calculations)
}
