% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_functions.R
\name{create.temp.vec}
\alias{create.temp.vec}
\title{Create a temperature vector}
\usage{
create.temp.vec(xy = xy, datemin = Sys.Date() - 365,
  datemax = Sys.Date())
}
\arguments{
\item{xy}{Coordinates (lat-long) for site to extract data from}

\item{datemin}{Earliest date to go back to get recrods for}

\item{datemax}{Latest date to get records for (default=current day)}
}
\description{
Downloads Australian Bureau of Meterology data and formats into a data.frame for use in development models
This function uses the 3 nearest stations to patch in missing data. Uses the the \code{bomrang} package
}
