% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_functions.R
\name{create_prec_vec}
\alias{create_prec_vec}
\title{Create a precipitation vector}
\usage{
create_prec_vec(xy = xy, datemin = Sys.Date() - 365,
  datemax = Sys.Date())
}
\arguments{
\item{xy}{Coordinates (lat-long) for site to extract data from}

\item{datemin}{Earliest date to go back to get recrods for}

\item{datemax}{Latest date to get records for (default=current day)}
}
\value{
A formatted data.frame with hourly interpolated precipitation data and the dates.
}
\description{
Downloads Australian Bureau of Meterology data and formats into a data.frame for use in development models
Uses the the \code{bomrang} package to obtain the data.
}
