% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_functions.R
\name{create_Ta}
\alias{create_Ta}
\title{Average Temp Raster}
\usage{
create_Ta(x, y, ext)
}
\arguments{
\item{x}{\code{RasterLayer} or \code{RasterBrick} object (e.g. Tmin)}

\item{y}{\code{RasterLayer} or \code{RasterBrick} object (e.g. Tmin)}

\item{ext}{\code{extent} object or vector with coordinates of extent}
}
\value{
\code{RasterLayer} or \code{RasterBrick} object with averaged temperatures
}
\description{
Simple function to create an average Temp layer from Tmin and Tmax layers or Bricks.
}
