% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earwig_dd.R
\name{earwig_dd}
\alias{earwig_dd}
\title{Earwig day degree simulations}
\usage{
earwig_dd(
  temps = temps,
  begin.date = Sys.Date(),
  date.error = 2,
  one.sided = TRUE,
  reps = 10,
  egg = TRUE
)
}
\arguments{
\item{temps}{Dataframe generated from \code{day_degree_simple}}

\item{begin.date}{Date to begin model runs}

\item{date.error}{How many days +/- of begin.date to sample from}

\item{one.sided}{True/False flag for if dates can only be selected BEFORE observation}

\item{reps}{Number or replicates}

\item{egg}{True/False flag of whether to include egg development}
}
\value{
A formatted data.frame with model outputs for each replicate
}
\description{
This is an example of how to run a number of day degree models with error around the parameters.
The names for the stages in the output are one stage along from the model names, this is because the model reflects that stage at end of development, whereas the model output reflects what stage would be observable.
Moult 4 is an adult earwig, so that is the stage observable following the completion of the model.
}
