% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SILO_data.R
\name{extract_SILO_data}
\alias{extract_SILO_data}
\title{Create Temperature Vectors from SILO data}
\usage{
extract_SILO_data(
  silodir = "/Users/hil32c/Documents/SILO_data/",
  site_list = as.data.frame(cbind(Site = "Test", x = -33.2833, y = 149.1)),
  datemin = "2015-06-30",
  datemax = "2021-03-10"
)
}
\arguments{
\item{silodir}{Directory holding the SILO netcdf files, both tmin and tmax for relevant years}

\item{datemin}{Date to begin extraction from}

\item{datemax}{Date to end extraction from}

\item{xy}{Coordinates of sites to extract data from}
}
\value{
A dataframe of Tmin and Tmax values for a given list of lat-longs
}
\description{
Uses the `{terra}` package to extract temperature vectors over multiple years from the gridded SILO data.
}
\author{
Matt Hill
}
