% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPFS.R
\name{GPFS_model}
\alias{GPFS_model}
\title{Generic Pest Forecast System (GPFS)}
\usage{
GPFS_model(weather.df = weather.df, Tmin = 13.3, Topt1 = 24,
  Topt2 = 34, Tmax = 41, Tmht = 33, Tmlt = 13.3, beta1 = 1101.7,
  beta2 = -49892, beta3 = -162.9, beta4 = 25.9595, beta5 = -0.4959,
  beta6 = 0, beta7 = 50.4, beta8 = 0.008, precipitation = FALSE)
}
\arguments{
\item{Tmin}{Temperature minimum}

\item{Topt1}{Low optimum temperature}

\item{Topt2}{High optimum temperature}

\item{Tmax}{Temperature maximum}

\item{Tmht}{Threshold - High temperature mortality}

\item{Tmlt}{Threshold - Low temperature mortality}

\item{beta1}{Constant for low temperature mortality}

\item{beta2}{Coefficient for mlt}

\item{beta3}{Coefficient for mlt}

\item{beta4}{Constant for mht}

\item{beta5}{Coefficient of degree one for mht}

\item{beta6}{Coefficient of degree two for mht}

\item{beta7}{Constant for Mwsm}

\item{beta8}{Reciprocal of number of hours required to reach maximum population for \code{P(n)}}

\item{precipitation}{T/F flag for whether to include the moisture parameters in the model}

\item{Mwsm}{Wet soil moisture mortality}
}
\value{
A data.frame with 7 columns showing parameters for Ta, P, d, mlt, mht and mwsm for hourly time steps over the data provided
}
\description{
Model is described in Hong et al. (2015). The population index each hour is based on the sum of a development rate (scaled by
the number of generations to reach maximum population) while removing the proportions
of the population killed by high and low temperatures and by wet soil moisture.
}
\details{
\code{P[n] <- (P[n-1] + (beta8 * d)) * (1-mlt) * (1-mht)}
}
\references{
Hong, S. C., Magarey, R., Borchert, D. M., Vargas, R. I., & Souder, S. (2015). Site-specific temporal and spatial validation of a generic plant pest forecast system with observations of Bactrocera dorsalis (oriental fruit fly). NeoBiota, 27, 37–67. https://doi.org/10.3897/neobiota.27.5177
}
\author{
Matt Hill
}
