% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_functions.R
\name{hourly.interpolate}
\alias{hourly.interpolate}
\title{Hourly temperature interpolation function}
\usage{
hourly.interpolate(climate = climate)
}
\arguments{
\item{climate}{A dataframe with columns of: (Date, Latitude, Longitude, Tmin, Tmax)}
}
\value{
An expanded data.frame of \code{climate} with 1 row per hour of each day provided, with corresponding temperature value
}
\description{
This function uses the \code{suncalc} package and a vector of dates to calculate solar noon and sunrise
Adjusts the sine coefficients to the location information, and from there assumes Tmin is achieved at
sunrise-1 hour and Tmax is achieved at solar noon + 2 hours.
Original calculations are found in chpater 2 of Campbell & Norman (1998)
}
\references{
Campbell, G. S., & Norman, J. M. (1998). An Introduction to Environmental Biophysics. https://doi.org/10.1007/978-1-4612-1626-1
}
\author{
Madeleine Barton

Matt Hill
}
