% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medfly_dd.R
\name{medfly_dd}
\alias{medfly_dd}
\title{Medfly day degree model}
\usage{
medfly_dd(temps = temps, begin.date = Sys.Date())
}
\arguments{
\item{temps}{Dataframe generated from \code{day_degree_simple}}

\item{begin.date}{Date to begin model runs}
}
\value{
A formatted data.frame with model outputs for each replicate
}
\description{
Description to come....
}
