% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_functions.R
\name{overwinter}
\alias{overwinter}
\title{Simple overwintering function}
\usage{
overwinter(tempvec = tempvec, datevec = datevec)
}
\description{
Switch for overwintering. This calculates cumulative days less than 18 degrees (5 needed to enter)
Currently calculated on the Ta (Average temperature for the day - based on the interpolated temp calculations)
}
