% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_functions.R
\name{overwinter_raster}
\alias{overwinter_raster}
\title{Simple overwintering function - raster based}
\usage{
overwinter_raster(brick_name = Ta_range, t.thresh = 18, n.days = 5)
}
\arguments{
\item{brick_name}{The \code{RasterBrick} object e.g. Tmin}

\item{t.thresh}{Threshold temperature in degrees}

\item{n.days}{Consecutive days needed at threshold to change condition}
}
\value{
\code{RasterLayer} object showing total number of overwintering days for the period analysed
}
\description{
Calculates number of overwintering days:
these are triggered by X consecutive days below a threshold temperature / broken by X days above
}
\author{
Matt Hill
}
