% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_functions.R
\name{overwinter_vector}
\alias{overwinter_vector}
\title{Simple overwintering function - vector based}
\usage{
overwinter_vector(tempvec = tempvec, datevec = datevec)
}
\description{
Switch for overwintering. This calculates cumulative days less than 18 degrees (5 needed to enter)
Currently calculated on the Ta (Average temperature for the day - based on the interpolated temp calculations)
}
