% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_functions.R
\name{spatial_degree_day}
\alias{spatial_degree_day}
\title{Cumulative degree day function - raster based}
\usage{
spatial_degree_day(
  Tmin = Tmin,
  Tmax = Tmax,
  tlow = 0,
  thigh = 100,
  len.layers = 365
)
}
\arguments{
\item{Tmin}{A \code{RasterBrick} object of minimum temperatures}

\item{Tmax}{A \code{RasterBrick} object of maximum temperatures}

\item{tlow}{Low temperature threshold}

\item{thigh}{High temperature threshold}

\item{len.layers}{How long (in days) to run the analysis for}
}
\value{
\code{RasterStack} object showing cumulative degree days for a given cell
}
\description{
Calculates cumulative degree days for RasterBrick objects
these are triggered by X consecutive days below a threshold temperature / broken by X days above
}
\author{
Matt Hill
}
