% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tephdev_vis.R
\name{TephDev_vis}
\alias{TephDev_vis}
\title{Shiny app to demonstrate model and data integration}
\usage{
TephDev_vis(silo_dir = silo_dir)
}
\value{
A Shiny app instance
}
\description{
In development!
}
\details{
Download the SILO data
}
\author{
Matt Hill
}
