% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yonow.R
\name{yonow_model}
\alias{yonow_model}
\title{Yonow et al. (2004) model for Queensland Fruit Fly development}
\usage{
yonow_model(tempvec = tempvec, datevec = datevec,
  start.date = start.date, ep1 = 0.0382, ep2 = 0.4229,
  lp1 = 0.0061, lp2 = 0.0609, pp1 = 0.0061, pp2 = 0.068,
  tf1 = 0.0108, tf2 = 0.133, ebt = 11.1, lbt = 10, pbt = 11.2,
  tfbt = 12.3, reverse = FALSE, adult = FALSE)
}
\arguments{
\item{tempvec}{A vector of temperatures (preferably daily average temp (Ta))}

\item{datevec}{A vector of dates that is the same length as \code{tempvec}}

\item{start.date}{Initialisation date, in YYYY-MM-DD format}

\item{ep1}{Egg development parameters}

\item{ep2}{Egg development parameters}

\item{lp1}{Larvae development parameters}

\item{lp2}{Larvae development parameters}

\item{pp1}{Pupae development parameters}

\item{pp2}{Pupae development parameters}

\item{tf1}{Teneral female development parameters}

\item{tf2}{Teneral female development parameters}

\item{ebt}{Egg base temperature}

\item{lbt}{Larval base temperature}

\item{pbt}{Pupal base temperature}

\item{tfbt}{Teneral female base temperature}

\item{reverse}{Run the model in reverse}

\item{adult}{Include / exclude the teneral female development stage}
}
\value{
A dataframe of daily dates, temperatures, insect development stages, and percent development
}
\description{
Function to run the Yonow et al. (2004) model on a temp and date vector. The initial parameters are for the
Queensland Fruit Fly (Bactrocera tryoni).
}
\references{
Yonow, T, M.P Zalucki, R.W Sutherst, B.C Dominiak, G.F Maywald, D.A Maelzer, and D.J Kriticos. “Modelling the Population Dynamics of the Queensland Fruit Fly, Bactrocera (Dacus) Tryoni: A Cohort-Based Approach Incorporating the Effects of Weather.” Ecological Modelling 173, no. 1 (March 2004): 9–30. https://doi.org/10.1016/S0304-3800(03)00306-5.
}
\author{
Matt Hill
}
