% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_functions.R
\name{day_degree_simple}
\alias{day_degree_simple}
\title{Simple day degree model}
\usage{
day_degree_simple(tlow = 4.5, thigh = 100, tempvec = tempvec,
  start.date = "2015-08-31")
}
\description{
Switch for overwintering. This calculates cumulative days less than 18 degrees (5 needed to enter)
Currently calculated on the Ta (Average temperature for the day - based on the interpolated temp calculations)
}
